+++ frontend/src/metabase/common/components/DatePicker/RelativeDatePicker/utils.ts
@@ -1,3 +1,5 @@
+import * as Lib from "metabase-lib";
+import { DATE_PICKER_TRUNCATION_UNITS } from "../constants";
 import type { RelativeDatePickerValue } from "../types";
 import { DEFAULT_VALUE } from "./constants";
 import type { IntervalDirection, RelativeDateIntervalValue } from "./types";
@@ -32,4 +34,27 @@ export function setDirection(
     case "next":
       return { ...valueOrDefault, value: Math.abs(valueOrDefault.value) };
   }
+}
+
+export function getInterval(value: RelativeDateIntervalValue): number {
+  return Math.abs(value.value);
+}
+
+export function setInterval(
+  value: RelativeDateIntervalValue,
+  interval: number,
+): RelativeDateIntervalValue {
+  const sign = Math.sign(value.value);
+
+  return {
+    ...value,
+    value: Math.max(Math.abs(Math.floor(interval)), 1) * sign,
+  };
+}
+
+export function getUnitOptions(interval: number) {
+  return DATE_PICKER_TRUNCATION_UNITS.map(unit => ({
+    value: unit,
+    label: Lib.describeTemporalUnit(unit, interval).toLowerCase(),
+  }));
 }
